+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Default drop down selection in Academics panel [#1726](https://github.com/rokwire/illinois-app/issues/1726).
 
 ## [4.0.28] - 2022-06-13
 ### Added